requirements.txt
Flask==1.1.2
locustio==0.999
pytest==7.1.1
pytest==7.3.1
